Renamed packages to follow standards
