+++ pom.xml
@@ -4,8 +4,8 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>ch.heigvd.SiTR</groupId>
-  <artifactId>SiTR</artifactId>
+  <groupId>ch.heigvd.sitr</groupId>
+  <artifactId>sitr</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
