+++ src/main/java/ch/heigvd/sitr/Vehicle.java
@@ -1,4 +1,4 @@
-package ch.heigvd.SiTR;
+package ch.heigvd.sitr;
 
 import lombok.Getter;
 import lombok.Setter;
