+++ src/main/java/ch/heigvd/sitr/gui/SettingsWindow.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ch.heigvd.SiTR.Viewer.SettingsWindow">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ch.heigvd.sitr.gui.SettingsWindow">
   <grid id="27dc6" binding="SettingsPanel" layout-manager="GridLayoutManager" row-count="10" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
