+++ src/test/java/ch/heigvd/sitr/AppTest.java
@@ -1,4 +1,4 @@
-package ch.heigvd.SiTR;
+package ch.heigvd.sitr;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
