+++ src/test/java/ch/heigvd/sitr/VehicleTest.java
@@ -1,8 +1,6 @@
-package ch.heigvd.SiTR;
+package ch.heigvd.sitr;
 
-import junit.framework.Test;
 import junit.framework.TestCase;
-import junit.framework.TestSuite;
 
 /**
  * Unit test for Vehicle.
